compacting signal concept
